+++ lib/params/message_list_params.g.dart
@@ -8,6 +8,14 @@ part of 'message_list_params.dart';
 
 MessageListParams _$MessageListParamsFromJson(Map<String, dynamic> json) {
   return MessageListParams()
+    ..includeMetaArray = json['with_sorted_meta_array'] as bool
+    ..includeReactions = json['include_reactions'] as bool
+    ..includeParentMessageText = json['include_parent_message_text'] as bool?
+    ..includeReplies = json['include_replies'] as bool?
+    ..includeThreadInfo = json['include_thread_info'] as bool
+    ..includeParentMessageInfo = json['include_parent_message_info'] as bool
+    ..replyType =
+        _$enumDecodeNullable(_$ReplyTypeEnumMap, json['include_reply_type'])
     ..previousResultSize = json['prev_limit'] as int
     ..nextResultSize = json['next_limit'] as int
     ..isInclusive = json['include'] as bool
@@ -20,19 +28,19 @@ MessageListParams _$MessageListParamsFromJson(Map<String, dynamic> json) {
         .toList()
     ..senderIds =
         (json['sender_ids'] as List<dynamic>?)?.map((e) => e as String).toList()
-    ..includeMetaArray = json['with_sorted_meta_array'] as bool
-    ..includeReactions = json['include_reactions'] as bool
-    ..includeParentMessageText = json['include_parent_message_text'] as bool
-    ..includeReplies = json['include_replies'] as bool
-    ..includeThreadInfo = json['include_thread_info'] as bool
     ..showSubChannelMessagesOnly =
-        json['show_sub_channel_messages_only'] as bool
-    ..includeParentMessageInfo = json['include_parent_message_info'] as bool
-    ..replyType = _$enumDecode(_$ReplyTypeEnumMap, json['include_reply_type']);
+        json['show_sub_channel_messages_only'] as bool;
 }
 
 Map<String, dynamic> _$MessageListParamsToJson(MessageListParams instance) =>
     <String, dynamic>{
+      'with_sorted_meta_array': instance.includeMetaArray,
+      'include_reactions': instance.includeReactions,
+      'include_parent_message_text': instance.includeParentMessageText,
+      'include_replies': instance.includeReplies,
+      'include_thread_info': instance.includeThreadInfo,
+      'include_parent_message_info': instance.includeParentMessageInfo,
+      'include_reply_type': _$ReplyTypeEnumMap[instance.replyType],
       'prev_limit': instance.previousResultSize,
       'next_limit': instance.nextResultSize,
       'include': instance.isInclusive,
@@ -41,14 +49,7 @@ Map<String, dynamic> _$MessageListParamsToJson(MessageListParams instance) =>
       'custom_type': instance.customType,
       'custom_types': instance.customTypes,
       'sender_ids': instance.senderIds,
-      'with_sorted_meta_array': instance.includeMetaArray,
-      'include_reactions': instance.includeReactions,
-      'include_parent_message_text': instance.includeParentMessageText,
-      'include_replies': instance.includeReplies,
-      'include_thread_info': instance.includeThreadInfo,
       'show_sub_channel_messages_only': instance.showSubChannelMessagesOnly,
-      'include_parent_message_info': instance.includeParentMessageInfo,
-      'include_reply_type': _$ReplyTypeEnumMap[instance.replyType],
     };
 
 K _$enumDecode<K, V>(
@@ -77,15 +78,26 @@ K _$enumDecode<K, V>(
   ).key;
 }
 
-const _$MessageTypeFilterEnumMap = {
-  MessageTypeFilter.all: null,
-  MessageTypeFilter.user: 'MESG',
-  MessageTypeFilter.file: 'FILE',
-  MessageTypeFilter.admin: 'ADMN',
-};
+K? _$enumDecodeNullable<K, V>(
+  Map<K, V> enumValues,
+  dynamic source, {
+  K? unknownValue,
+}) {
+  if (source == null) {
+    return null;
+  }
+  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
+}
 
 const _$ReplyTypeEnumMap = {
   ReplyType.none: 'NONE',
   ReplyType.all: 'ALL',
   ReplyType.only_reply_to_channel: 'ONLY_REPLY_TO_CHANNEL',
+};
+
+const _$MessageTypeFilterEnumMap = {
+  MessageTypeFilter.all: null,
+  MessageTypeFilter.user: 'MESG',
+  MessageTypeFilter.file: 'FILE',
+  MessageTypeFilter.admin: 'ADMN',
 };