+++ lib/params/message_retrieval_params.g.dart
@@ -13,28 +13,29 @@ MessageRetrievalParams _$MessageRetrievalParamsFromJson(
     channelUrl: json['channel_url'] as String,
     messageId: json['message_id'] as int,
   )
-    ..includeReactions = json['include_reactions'] as bool
     ..includeMetaArray = json['with_sorted_meta_array'] as bool
-    ..includeParentMessageText = json['include_parent_message_text'] as bool
+    ..includeReactions = json['include_reactions'] as bool
+    ..includeParentMessageText = json['include_parent_message_text'] as bool?
+    ..includeReplies = json['include_replies'] as bool?
     ..includeThreadInfo = json['include_thread_info'] as bool
-    ..includeReplies = json['include_replies'] as bool
     ..includeParentMessageInfo = json['include_parent_message_info'] as bool
-    ..replyToChannel = json['reply_to_channel'] as bool;
+    ..replyType =
+        _$enumDecodeNullable(_$ReplyTypeEnumMap, json['include_reply_type']);
 }
 
 Map<String, dynamic> _$MessageRetrievalParamsToJson(
         MessageRetrievalParams instance) =>
     <String, dynamic>{
+      'with_sorted_meta_array': instance.includeMetaArray,
+      'include_reactions': instance.includeReactions,
+      'include_parent_message_text': instance.includeParentMessageText,
+      'include_replies': instance.includeReplies,
+      'include_thread_info': instance.includeThreadInfo,
+      'include_parent_message_info': instance.includeParentMessageInfo,
+      'include_reply_type': _$ReplyTypeEnumMap[instance.replyType],
       'channel_url': instance.channelUrl,
       'channel_type': _$ChannelTypeEnumMap[instance.channelType],
       'message_id': instance.messageId,
-      'include_reactions': instance.includeReactions,
-      'with_sorted_meta_array': instance.includeMetaArray,
-      'include_parent_message_text': instance.includeParentMessageText,
-      'include_thread_info': instance.includeThreadInfo,
-      'include_replies': instance.includeReplies,
-      'include_parent_message_info': instance.includeParentMessageInfo,
-      'reply_to_channel': instance.replyToChannel,
     };
 
 K _$enumDecode<K, V>(
@@ -66,4 +67,21 @@ K _$enumDecode<K, V>(
 const _$ChannelTypeEnumMap = {
   ChannelType.group: 'group',
   ChannelType.open: 'open',
+};
+
+K? _$enumDecodeNullable<K, V>(
+  Map<K, V> enumValues,
+  dynamic source, {
+  K? unknownValue,
+}) {
+  if (source == null) {
+    return null;
+  }
+  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
+}
+
+const _$ReplyTypeEnumMap = {
+  ReplyType.none: 'NONE',
+  ReplyType.all: 'ALL',
+  ReplyType.only_reply_to_channel: 'ONLY_REPLY_TO_CHANNEL',
 };